pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "codepilot"
version = "0.1.0"
description = "CLI coding assistant powered by Qwen3 Coder"
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.27",
    "pydantic>=2.0",
    "rich>=13.0",
    "click>=8.0",
    "pyyaml>=6.0",
]

[project.scripts]
codepilot = "codepilot.cli.app:cli"

[tool.setuptools.packages.find]
include = ["codepilot*"]
